requirements.txt
supervision
ultralytics
openvino
opencv-python
flask
pathlib
gunicorn
flask-mqtt
numpy
pillow